feat(v6): description of the v6 encryption format
Note that I described this format as the default one even though it is not the
default one yet.
I avoided mentioning the option capable of manually setting the
padding step. I feel like there is no right place to write about it as
it consists of an API behavior.